complaints_ai/requirements.txt
polars
sqlalchemy
pymysql
streamlit
pandas
numpy
scipy
python-dotenv
pyyaml
